+++ pmd-test/src/main/java/net/sourceforge/pmd/testframework/TestDescriptor.java
@@ -13,11 +13,15 @@ import org.junit.Ignore;
 
 import net.sourceforge.pmd.Rule;
 import net.sourceforge.pmd.lang.LanguageVersion;
+import net.sourceforge.pmd.test.schema.RuleTestDescriptor;
 
 /**
  * Stores the information required to run a complete test.
+ *
+ * @deprecated Use {@link RuleTestDescriptor} instead
  */
 @Ignore("this is not a unit test")
+@Deprecated
 public class TestDescriptor {
     private Rule rule;
     private Properties properties;
@@ -44,12 +48,28 @@ public class TestDescriptor {
     }
 
     public TestDescriptor(String code, String description, int numberOfProblemsExpected, Rule rule,
-            LanguageVersion languageVersion) {
+                          LanguageVersion languageVersion) {
         this.rule = rule;
         this.code = code;
         this.description = description;
         this.numberOfProblemsExpected = numberOfProblemsExpected;
         this.languageVersion = languageVersion;
+    }
+
+    // for compatibility
+    TestDescriptor(RuleTestDescriptor td, String absoluteUriToTestXmlFile) {
+        this.rule = td.getRule();
+        this.code = td.getCode();
+        this.description = td.getDescription();
+        this.numberOfProblemsExpected = td.getExpectedProblems();
+        this.expectedLineNumbers = td.getExpectedLineNumbers();
+        this.expectedMessages = td.getExpectedMessages();
+        this.isRegressionTest = !td.isDisabled();
+        this.numberInDocument = td.getIndex();
+        this.properties = td.getProperties();
+        this.languageVersion = td.getLanguageVersion();
+        this.numberInDocument = td.getIndex();
+        this.setTestSourceUri(absoluteUriToTestXmlFile, td.getLineNumber());
     }
 
     public int getNumberInDocument() {